Extract duplicated getFileType() and FileIcon from ReadDisplay and WriteDisplay
Move the shared getFileType() function to src/lib/file-types.ts and the
FileIcon SVG component to src/components/messages/FileIcon.tsx. Both
ReadDisplay and WriteDisplay now import from these shared modules instead
of maintaining identical copies.

The FileIcon component accepts a variant prop ("read" or "write") to
preserve the slightly different SVG paths used by each display.

Fixes #208
