+++ src/components/messages/FileIcon.tsx
@@ -1,0 +1,30 @@
+/**
+ * File icon SVG component used in tool displays.
+ *
+ * variant="read" shows a plain document icon.
+ * variant="write" shows a document icon with lines inside it.
+ */
+export function FileIcon({
+  className,
+  variant = 'read',
+}: {
+  className?: string;
+  variant?: 'read' | 'write';
+}) {
+  const path =
+    variant === 'write'
+      ? 'M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m0 12.75h7.5m-7.5 3H12M10.5 2.25H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z'
+      : 'M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m2.25 0H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z';
+
+  return (
+    <svg
+      className={`w-4 h-4 text-muted-foreground flex-shrink-0 ${className ?? ''}`}
+      fill="none"
+      viewBox="0 0 24 24"
+      stroke="currentColor"
+      strokeWidth={1.5}
+    >
+      <path strokeLinecap="round" strokeLinejoin="round" d={path} />
+    </svg>
+  );
+}
