+++ src/components/messages/ReadDisplay.tsx
@@ -2,6 +2,8 @@
 
 import { useMemo } from 'react';
 import { Badge } from '@/components/ui/badge';
+import { getFileType } from '@/lib/file-types';
+import { FileIcon } from './FileIcon';
 import { ToolDisplayWrapper } from './ToolDisplayWrapper';
 import type { ToolCall } from './types';
 
@@ -61,67 +63,6 @@ function parseReadOutput(output: string): ParsedLine[] {
   }
 
   return result;
-}
-
-/**
- * Detect file type from extension for syntax highlighting hints.
- */
-function getFileType(filePath: string): string {
-  const ext = filePath.split('.').pop()?.toLowerCase() ?? '';
-  const typeMap: Record<string, string> = {
-    ts: 'typescript',
-    tsx: 'typescript',
-    js: 'javascript',
-    jsx: 'javascript',
-    py: 'python',
-    rb: 'ruby',
-    rs: 'rust',
-    go: 'go',
-    java: 'java',
-    kt: 'kotlin',
-    swift: 'swift',
-    c: 'c',
-    cpp: 'cpp',
-    h: 'c',
-    hpp: 'cpp',
-    css: 'css',
-    scss: 'scss',
-    less: 'less',
-    html: 'html',
-    xml: 'xml',
-    json: 'json',
-    yaml: 'yaml',
-    yml: 'yaml',
-    md: 'markdown',
-    sql: 'sql',
-    sh: 'shell',
-    bash: 'shell',
-    zsh: 'shell',
-    dockerfile: 'docker',
-    prisma: 'prisma',
-  };
-  return typeMap[ext] ?? 'text';
-}
-
-/**
- * FileIcon component for Read tool display
- */
-function FileIcon({ className }: { className?: string }) {
-  return (
-    <svg
-      className={`w-4 h-4 text-muted-foreground flex-shrink-0 ${className ?? ''}`}
-      fill="none"
-      viewBox="0 0 24 24"
-      stroke="currentColor"
-      strokeWidth={1.5}
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m2.25 0H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z"
-      />
-    </svg>
-  );
 }
 
 /**
